Реализовать передачу зависимостей при выполнении ядер

+++ ocl_project/BicgStabHost.cs
@@ -3,8 +3,8 @@ using System.Diagnostics;
 
 using static Solvers.Shared;
 
-using Real = double;
-using VectorReal = double[];
+using Real = float;
+using VectorReal = float[];
 
 namespace Solvers.Host
 {
